requirements.txt
# SCLC Chemo-IO Resistance Pipeline Dependencies
# Install with: pip install -r requirements.txt

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.9.0

# Gene set enrichment
gseapy>=1.0.0

# Network analysis
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# GEO/SRA data access
GEOparse>=2.0.0
pysradb>=2.0.0

# API requests
requests>=2.28.0
tenacity>=8.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# File handling
openpyxl>=3.1.0
xlrd>=2.0.0
h5py>=3.8.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.2.0

# Survival analysis (optional)
lifelines>=0.27.0

# Progress bars
tqdm>=4.65.0

# Type hints (development)
typing-extensions>=4.5.0
